move information_panel.rs
