+++ src/data/mod.rs
@@ -6,7 +6,4 @@ pub use disks::{Disk, Disks};
 
 mod user_directories;
 
-/// Information panel showing the preview and metadata of the selected item
-pub mod information_panel;
-
 pub use user_directories::UserDirectories;
