+++ src/information_panel.rs
